add mini-css-extract-plugin to generate significantly smaller bundles which allows enableJs to work for percy
